requirements.txt
nltk==3.7
scikit_learn==1.1.1
pyTelegramBotAPI
python-dotenv